Drought Prediction Web Application
This is the web application to receive rainfall information and calculate.
When filling the blank in the webpage, the drop down box could change dynamically by the previous choose. As there are many detect stations, different station has different time range to be filled with rainfall information. When station is choosed in drop down box on the webpage, it will automatically search compliant year and month in DB via JavaScript. 
The algorithms written in Matlab by my cooperator who studies in Environmental Science have been packaged into jar flies to be used in backend. Finally, the calculate result is stored in MySQL.